+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 All notable changes to homebridge-dummy will be documented in this file.
 
-## 1.5.0-beta.0 (2025-10-??)
+## 1.5.0-beta.0 (2025-10-20)
 
 ### Added
+- [Value Transformers](https://github.com/mpatfield/homebridge-easy-mqtt#history) (Thank you [@nehmeroumani](https://github.com/sponsors/nehmeroumani) for the code contribution!)
 - `Doorbell` accessory type
 - History and additional characteristics for for [Eve app](https://apps.apple.com/us/app/eve-for-matter-homekit/id917695792) users ([documentation](https://github.com/mpatfield/homebridge-easy-mqtt#history))
     - Opened/closed history, and times opened count with option to reset for `ContactSensor`